requirements.txt
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
streamlit==1.28.1
requests==2.31.0
python-dotenv==1.0.0
imbalanced-learn==0.11.0
jupyter==1.0.0
ipykernel==6.25.2
dask[complete]==2023.8.1
folium==0.14.0